requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
Brotli==1.0.9
cffi==1.15.1
channels==2.4.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
dash==2.6.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dj-database-url==0.5.0
Django==3.2.14
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-plotly-dash==2.0.0
dpd-components==0.1.0
Flask==2.1.3
Flask-Compress==1.12
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
pandas==1.4.3
plotly==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
Werkzeug==2.0.3
whitenoise==6.2.0
zipp==3.8.1
zope.interface==5.4.0
